+++ packages/browser/src/index.ts
@@ -6,5 +6,6 @@ export * from "./keyboard";
 export * from "./mouse";
 export * from "./runner";
 export * from "./video";
+export * from "./download";
 
 export const Name = "@a2-ts-utils/browser";